+++ packages/gestalt/src/Sticky.js
@@ -3,6 +3,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import layout from './Layout.css';
+import { type Indexable, FixedZIndex } from './zIndex.js';
 
 type Threshold =
   | {| top: number | string |}
@@ -20,19 +21,28 @@ type Threshold =
 
 type Props = {|
   children: React.Node,
+  zIndex?: Indexable,
   dangerouslySetZIndex?: {| __zIndex: number |},
   ...Threshold,
 |};
 
+const DEFAULT_ZINDEX = new FixedZIndex(1);
+
 export default function Sticky(props: Props): React.Node {
-  const { dangerouslySetZIndex = { __zIndex: 1 }, children } = props;
+  const { dangerouslySetZIndex, children } = props;
+  const zIndex =
+    props.zIndex ||
+    (dangerouslySetZIndex &&
+    Object.prototype.hasOwnProperty.call(dangerouslySetZIndex, '__zIndex')
+      ? // eslint-disable-next-line no-underscore-dangle
+        new FixedZIndex(dangerouslySetZIndex.__zIndex)
+      : DEFAULT_ZINDEX);
   const style = {
     top: props.top != null ? props.top : undefined,
     left: props.left != null ? props.left : undefined,
     right: props.right != null ? props.right : undefined,
     bottom: props.bottom != null ? props.bottom : undefined,
-    // eslint-disable-next-line no-underscore-dangle
-    zIndex: dangerouslySetZIndex.__zIndex,
+    zIndex: zIndex.index(),
   };
   return (
     <div className={layout.sticky} style={style}>
@@ -50,4 +60,6 @@ Sticky.propTypes = {
   left: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   bottom: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   right: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  // eslint-disable-next-line react/forbid-prop-types
+  zIndex: PropTypes.any,
 };