+++ packages/gestalt/src/index.js
@@ -12,6 +12,7 @@ import Column from './Column.js';
 import Container from './Container.js';
 import Divider from './Divider.js';
 import Flyout from './Flyout.js';
+import GestaltProvider from './GestaltProvider.js';
 import GroupAvatar from './GroupAvatar.js';
 import Heading from './Heading.js';
 import Icon from './Icon.js';
@@ -47,6 +48,8 @@ import TextField from './TextField.js';
 import Toast from './Toast.js';
 import Tooltip from './Tooltip.js';
 import Video from './Video.js';
+import { FixedZIndex, CompositeZIndex } from './zIndex.js';
+import { useTheme } from './contexts/Theme.js';
 
 export {
   Avatar,
@@ -59,9 +62,12 @@ export {
   Collage,
   Collection,
   Column,
+  CompositeZIndex,
   Container,
   Divider,
+  FixedZIndex,
   Flyout,
+  GestaltProvider,
   GroupAvatar,
   Heading,
   Icon,
@@ -96,5 +102,6 @@ export {
   TextField,
   Toast,
   Tooltip,
+  useTheme,
   Video,
 };